[#156913774] summary component optimization
